# backend/docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrueAdmin API",
    "version": "0.1.0",
    "description": "Generated from TrueAdmin controller attributes. Code annotations provide default interface metadata."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/admin/messages": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_list",
        "summary": "消息中心列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@list",
          "permission": "system:message:list",
          "permissionMode": "single",
          "permissions": [
            "system:message:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/messages/{kind}/{id}": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_detail",
        "summary": "消息详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@detail",
          "permission": "system:message:detail",
          "permissionMode": "single",
          "permissions": [
            "system:message:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/messages/archive": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_archive",
        "summary": "归档消息",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@archive",
          "permission": "system:message:update",
          "permissionMode": "single",
          "permissions": [
            "system:message:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/messages/read": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_markRead",
        "summary": "标记消息已读",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@markRead",
          "permission": "system:message:update",
          "permissionMode": "single",
          "permissions": [
            "system:message:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/messages/read-all": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_readAll",
        "summary": "全部标记已读",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@readAll",
          "permission": "system:message:update",
          "permissionMode": "single",
          "permissions": [
            "system:message:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/messages/restore": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_restore",
        "summary": "恢复归档消息",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@restore",
          "permission": "system:message:update",
          "permissionMode": "single",
          "permissions": [
            "system:message:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/messages/stream": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_messageStream",
        "summary": "消息实时同步",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@messageStream",
          "permission": "system:message:list",
          "permissionMode": "single",
          "permissions": [
            "system:message:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/messages/unread-count": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMessageController_unreadCount",
        "summary": "消息未读数量",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMessageController@unreadCount",
          "permission": "system:message:list",
          "permissionMode": "single",
          "permissions": [
            "system:message:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/organization/roles/{id}/menus": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_authorize",
        "summary": "角色菜单授权",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@authorize",
          "permission": "system:role:authorize",
          "permissionMode": "single",
          "permissions": [
            "system:role:authorize"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/system-config/menus": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMenuController_create",
        "summary": "新增菜单",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMenuController@create",
          "permission": "system:menu:create",
          "permissionMode": "single",
          "permissions": [
            "system:menu:create"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMenuController_list",
        "summary": "菜单列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMenuController@list",
          "permission": "system:menu:list",
          "permissionMode": "single",
          "permissions": [
            "system:menu:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/system-config/menus/{id}": {
      "delete": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMenuController_delete",
        "summary": "删除菜单",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMenuController@delete",
          "permission": "system:menu:delete",
          "permissionMode": "single",
          "permissions": [
            "system:menu:delete"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMenuController_detail",
        "summary": "菜单详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMenuController@detail",
          "permission": "system:menu:detail",
          "permissionMode": "single",
          "permissions": [
            "system:menu:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMenuController_update",
        "summary": "编辑菜单",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMenuController@update",
          "permission": "system:menu:update",
          "permissionMode": "single",
          "permissions": [
            "system:menu:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/system-config/menus/tree": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminMenuController_tree",
        "summary": "菜单树",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminMenuController@tree",
          "permission": "system:menu:list",
          "permissionMode": "single",
          "permissions": [
            "system:menu:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}